requirements/prod.txt
Django==3.0.4
gunicorn==20.0.4
django-crispy-forms==1.9.0
python3-logstash==0.4.80
djangorestframework==3.11.0
